refactor: DH-23209: WidgetLoaderUtils now exported from dashboard-core-plugins
- There are other cases where we just want to get the plugin for a given widget, and don't want to use a WidgetLoaderPlugin
- Export the logic so it can be reused
